% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_date_variables.R
\name{add_date_variables}
\alias{add_date_variables}
\title{Function to transform a parsed \code{date} variable into other date/time 
variables in preparation for modelling with the \strong{normalweatherr} 
package.}
\usage{
add_date_variables(df)
}
\arguments{
\item{df}{Data frame.}
}
\value{
Data frame.
}
\description{
\code{prepare_input_data} will check variable names, transform the parsed 
\code{date} variable into other variables, impute missing data, and make 
correct data types.
}
\author{
Stuart K. Grange
}
