% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_edges.R
\name{clip_edges}
\alias{clip_edges}
\title{Function to ``clip'' the edges of a time-series.}
\usage{
clip_edges(df, seconds = 31536000/2, filter = FALSE)
}
\arguments{
\item{df}{Data frame.}

\item{seconds}{Number of seconds to clip from start and end of time-series. 
The default is half a year.}

\item{filter}{Should the observations to be clipped be filtered from \code{df}
or invalidated with \code{NA}?}
}
\value{
Data frame.
}
\description{
Function to ``clip'' the edges of a time-series.
}
\author{
Stuart K. Grange
}
