% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseasonalise_stl.R
\name{deseasonalise_stl}
\alias{deseasonalise_stl}
\title{Function to decompose a monthly time series with loess.}
\usage{
deseasonalise_stl(df)
}
\arguments{
\item{df}{Data frame containing a monthly time series. \code{df} must contain
\code{"date"} and \code{"value"} variables.}
}
\value{
Numeric vector with the length of \code{nrow(df)}.
}
\description{
Function to decompose a monthly time series with loess.
}
\seealso{
\code{\link{stl}}
}
\author{
Stuart K. Grange
}
