% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorological_functions.R
\name{detect_breakpoints}
\alias{detect_breakpoints}
\title{Function to detect breakpoints in a data frame.}
\usage{
detect_breakpoints(df)
}
\arguments{
\item{df}{Data frame to detect breakpoints in.}
}
\value{
Data frame.
}
\description{
Function to detect breakpoints in a data frame.
}
\author{
Stuart K. Grange
}
