% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtest_prediction.R
\name{ggtest_prediction}
\alias{ggtest_prediction}
\title{Function to use the testing set in a \strong{normalweatherr_model} object and
plot the values and predicted values as a hex-plot.}
\usage{
ggtest_prediction(list_model)
}
\arguments{
\item{list_model}{A \strong{normalweatherr_model} object.}
}
\value{
Invisible, a plot.
}
\description{
Function to use the testing set in a \strong{normalweatherr_model} object and
plot the values and predicted values as a hex-plot.
}
\author{
Stuart K. Grange
}
