% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorological_functions.R
\name{normalise_for_meteorology}
\alias{normalise_for_meteorology}
\title{Function to normalise a concentration variable based on "average" 
meteorological conditions.}
\usage{
normalise_for_meteorology(list_model, df, variables = c("wd", "ws", "temp",
  "rh", "hour", "weekday", "week"), n = 100, replace = FALSE, output = NA)
}
\arguments{
\item{list_model}{A \code{randomForest} model object, typically created with
\code{\link{calculate_model}}.}

\item{df}{Data frame to use for prediction. Typically created with 
\code{\link{prepare_input_data}}.}

\item{variables}{Variable to randomly sample.}

\item{n}{Number of times to sample \code{df}.}

\item{replace}{Should \code{variables} be sampled with replacement?}

\item{output}{Directory to export the model object as an \code{.rds} file. 
If not used, the model will not be exported to disc.}
}
\value{
Data frame.
}
\description{
Function to normalise a concentration variable based on "average" 
meteorological conditions.
}
\author{
Stuart K. Grange
}
