% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_variables.R
\name{normalise_variables}
\alias{normalise_variables}
\title{Function to normalise for meteorology for many variables.}
\usage{
normalise_variables(df, variables, variables_explanatory, progress = "time")
}
\arguments{
\item{df}{Data frame containing a \code{date} variable.}

\item{variables}{Variables in \code{df} to normalise for meteorology.}

\item{variables_explanatory}{Variables to include in the model, \emph{i.e.} 
the predictors.}

\item{progress}{Type of progress bar to display.}
}
\value{
Named list with split input data, model performance data, and 
normalised data.
}
\description{
Function to normalise for meteorology for many variables.
}
\author{
Stuart K. Grange
}
