% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_using_test_set.R
\name{predict_using_test_set}
\alias{predict_using_test_set}
\title{Function to use a test set for predictions within a 
\strong{normalweatherr_model} object.}
\usage{
predict_using_test_set(list_model)
}
\arguments{
\item{A}{\strong{normalweatherr_model} object.}
}
\value{
Data frame.
}
\description{
Function to use a test set for predictions within a 
\strong{normalweatherr_model} object.
}
\author{
Stuart K. Grange
}
