% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input_data.R
\name{prepare_input_data}
\alias{prepare_input_data}
\title{Function to prepare meteorological and air quality data for modelling with
the \strong{normalweatherr} package.}
\usage{
prepare_input_data(df, impute = TRUE, fraction = 0.8)
}
\arguments{
\item{df}{Data frame containing \code{date} and \code{value} variables.}

\item{impute}{Should missing values be imputed?}

\item{fraction}{Fraction of observations to form the training set. Default is
\code{0.8} for an 80/20 \% split for training and testing sets.}
}
\value{
A named list containing two data frames: \code{training} and 
\code{testing} with the class \code{normalweatherr_data}.
}
\description{
\code{prepare_input_data} will check variable names, transform the parsed 
\code{date} variable into other variables, impute missing data, make correct
data types, and split the input data set into training and testing sets.
}
\details{
Use \code{set.seed} to ensure the splitting of the training and testing sets
is reproducable.
}
\examples{
\dontrun{

# Make modelling reproducable
set.seed(123)

# Prepare data for modelling
list_input_data <- prepare_input_data(data_swiss_daily)

}

}
\seealso{
\code{\link{set.seed}}, \code{\link{calculate_model}}, 
\code{\link{normalise_for_meteorology}}
}
\author{
Stuart K. Grange
}
