% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalweatherr_helpers.R
\name{register_cores}
\alias{register_cores}
\title{Function to register parallel backend.}
\usage{
register_cores(cores = NA)
}
\value{
Invisible.
}
\description{
Function to register parallel backend.
}
\author{
Stuart K. Grange
}
