% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_input_data.R
\name{split_input_data}
\alias{split_input_data}
\title{Function to split an input data frame into training and testing sets.}
\usage{
split_input_data(df, fraction = 0.8)
}
\arguments{
\item{df}{Data frame containing \code{date} and \code{value} variables. 
\code{df} will usually be prepared with \code{\link{add_date_variables}}.}

\item{fraction}{Fraction of observations to form the training set. Default is
\code{0.8} for an 80/20 \% split for training and testing sets.}
}
\value{
A named list containing two data frames: \code{training} and 
\code{testing} with the class \code{normalweatherr_data}.
}
\description{
Use \code{set.seed} to ensure the splitting of the training and testing sets
is reproducible.
}
\examples{
\dontrun{

# Make modelling reproducible
set.seed(123)

# Prepare data for modelling
list_input_data <- split_input_data(data_swiss_daily)

}

}
\seealso{
\code{\link{set.seed}}, \code{\link{add_date_variables}},
 \code{\link{calculate_model}}, \code{\link{normalise_for_meteorology}}
}
\author{
Stuart K. Grange
}
