% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theil_sen_trend_test.R
\name{theil_sen_trend_test}
\alias{theil_sen_trend_test}
\title{Function to test a time series with the Theil-Sen estimator.}
\usage{
theil_sen_trend_test(df, variable = "value", deseason = FALSE,
  auto_correlation = FALSE)
}
\arguments{
\item{df}{Data frame.}

\item{variable}{Variable name to test.}

\item{deseason}{Should the time series be deseaonsalised before the trend 
test is conducted?}

\item{auto_correlation}{Should auto correlation be considered in the estimates?}
}
\value{
Data frame.
}
\description{
Function to test a time series with the Theil-Sen estimator.
}
\seealso{
\code{\link{TheilSen}}
}
\author{
Stuart K. Grange
}
