% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theil_sen_trend_test.R
\name{theil_sen_trend_vector}
\alias{theil_sen_trend_vector}
\title{Function to calculate a trend vector for a time series.}
\usage{
theil_sen_trend_vector(df, variable = "value", deseason = FALSE,
  auto_correlation = FALSE, df_trend = NA)
}
\arguments{
\item{df}{Input data frame.}

\item{variable}{Variable to test.}

\item{deseason}{Should the time series be deseaonsalised before the trend 
test is conducted?}

\item{auto_correlation}{Should auto correlation be considered in the estimates?}

\item{df_trend}{An existing trend data frame from 
\code{\link{theil_sen_trend_test}}. An optional argument.}
}
\value{
Numeric vector with the length of \code{nrow(df)}.
}
\description{
Function to calculate a trend vector for a time series.
}
\seealso{
\code{\link{theil_sen_trend_test}}
}
\author{
Stuart K. Grange
}
