% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorological_functions.R
\name{calculate_model}
\alias{calculate_model}
\title{Function to model concentration based on meteorological and time variables.}
\usage{
calculate_model(list_input_data, variables = c("temp", "rh", "ws", "wd",
  "date_unix", "week", "weekday", "hour"), ntree = 200, mtry = 3,
  nodesize = 3, verbose = TRUE, output = NA)
}
\arguments{
\item{list_input_data}{\code{normalweatherr_data} list containing the prepared
training and testing sets.}

\item{variables}{Variables to include in the random forest model.}

\item{ntree}{Number of trees to grow for the random forest model. Set 
\code{ntree} to a smaller integer for testing.}

\item{mtry}{Number of variables randomly sampled for splitting the random 
forest decision tree.}

\item{nodesize}{Minimum size of terminal nodes for the random forest model.}

\item{verbose}{Should the random forest model print progress.}

\item{output}{Directory to export the model object as an \code{.rds} file. 
If not used, the model will not be exported to disc.}
}
\value{
Named list containing two data frames and a model object with the 
class \code{normalweatherr_model}.
}
\description{
Function to model concentration based on meteorological and time variables.
}
\author{
Stuart K. Grange
}
