% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_model.R
\name{calculate_model}
\alias{calculate_model}
\title{Function to model pollutant concentrations based on meteorological and time
variables.}
\usage{
calculate_model(list_input_data, variables, model = "rf", ntree = NA,
  mtry = 3, nodesize = 3, verbose = TRUE, output = NA)
}
\arguments{
\item{list_input_data}{\code{normalweatherr_data} list containing the prepared
training and testing sets; produced by \code{\link{split_input_data}}.}

\item{variables}{Variables to include in the model.}

\item{model}{Model type to use. Can be: \code{"rf"}, \code{"svm"}, \code{"gam"}, 
\code{"gbm"} for random forest, support vector machines, generalised additive
models, and generalised boosted regression models respectively.}

\item{ntree}{Number of trees to grow for \code{"rf"} or \code{"gbm"} models. 
Set \code{ntree} to a smaller integer for testing.}

\item{mtry}{Number of variables randomly sampled for splitting the \code{"rf"}
decision tree.}

\item{nodesize}{Minimum size of terminal nodes for the \code{"rf"} model.}

\item{verbose}{Should the models print progress if this is supported?}

\item{output}{File name to export the model object as an \code{.rds} file. 
If not used, the model will not be exported to disc. Directories will be 
created if necessary.}
}
\value{
Named list containing two data frames (contents of 
\code{list_input_data}) and a model object with the class 
\code{normalweatherr_model}.
}
\description{
To-do: Support for models other than random forest is pretty poor at the 
moment, this needs to be fixed.
}
\examples{
\dontrun{

# Calculate a random forest model
list_model <- calculate_model(
  list_input_data = list_input_data, 
  variables = variables, 
  model = "rf",
  ntree = 200,
  output = NA,
  mtry = 5,
  nodesize = 1,
  verbose = TRUE
)

}

}
\seealso{
\code{\link{split_input_data}}, 
\code{\link{normalise_for_meteorology}}
}
\author{
Stuart K. Grange
}
