% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_breakpoints.R
\name{detect_breakpoints}
\alias{detect_breakpoints}
\title{Function to detect breakpoints in a data frame.}
\usage{
detect_breakpoints(df, h = 0.15, n = NULL)
}
\arguments{
\item{df}{Data frame to detect breakpoints in.}

\item{h}{Minimal segment size either given as fraction relative to the sample 
size or as an integer giving the minimal number of observations in each 
segment.}

\item{n}{Number of breaks to detect. Default is maximum number allowed by
\code{h}.}
}
\value{
Data frame.
}
\description{
Function to detect breakpoints in a data frame.
}
\author{
Stuart K. Grange
}
