% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorological_functions.R
\name{prepare_input_data}
\alias{prepare_input_data}
\title{Function to prepare meteorological and air quality data for modelling.}
\usage{
prepare_input_data(df)
}
\arguments{
\item{df}{Data frame containing \code{date} and \code{value} variables.}
}
\value{
A named list containing two data frames with the class 
\code{normalweatherr_data}.
}
\description{
\code{prepare_input_data} will check variable names, transform the parsed 
\code{date} variable into other variables, impute missing data, and split
the data set into the training and testing sets.
}
\author{
Stuart K. Grange
}
